Moved menus ui to own package
